% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{load_janesick_ad}
\alias{load_janesick_ad}
\title{Load Janesick Breast Cancer Crop (AnnData)}
\usage{
load_janesick_ad()
}
\value{
An AnnData object (via reticulate).
}
\description{
Loads a 25000x25000 pixel crop of the 10x Genomics Xenium Human Breast
Cancer FFPE dataset (Rep 1).
}
\details{
Reference: Janesick et al., Nature Communications, 2023.
Source: https://www.10xgenomics.com/datasets/xenium-ffpe-human-breast-cancer-rep1
}
