% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{load_janesick_spe}
\alias{load_janesick_spe}
\title{Load Janesick Breast Cancer Crop (rds)}
\usage{
load_janesick_spe()
}
\value{
A SpatialExperiment object.
}
\description{
Loads a 25000x25000 pixel crop of the 10x Genomics Xenium Human Breast
Cancer FFPE dataset (Rep 1) as a SpatialExperiment object.
}
\details{
Reference: Janesick et al., Nature Communications, 2023.
Source: https://www.10xgenomics.com/datasets/xenium-ffpe-human-breast-cancer-rep1
}
