% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{novae_run}
\alias{novae_run}
\title{Perform a one-slide zero-short inference.}
\usage{
novae_run(spe, gpu = FALSE)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} object.}

\item{gpu}{Use a GPU accelerator.}
}
\value{
The embedding matrix computed by \code{novae}.
}
\description{
Perform a one-slide zero-short inference.
}
